refactor configuration functions iunto read and update modules, with a barrel for minimal refactor issues. Just to keep code maint easier
