+++ src/test/suite/surprise-me-on-startup.test.ts
@@ -9,14 +9,14 @@ import {
 } from '../../models';
 import { setupTestSuite, teardownTestSuite, setupTest } from './lib/setup-teardown-test-suite';
 import { executeCommand } from './lib/constants';
+import { applyInitialConfiguration } from '../../extension';
 import {
   getOriginalColorsForAllElements,
-  updateSurpriseMeOnStartup,
   updateAffectedElements,
+  updateSurpriseMeOnStartup,
 } from '../../configuration';
-import { applyInitialConfiguration } from '../../extension';
 
-suite('Surprise me on startup', () => {
+suite.only('Surprise me on startup', () => {
   let originalValues = <IPeacockSettings>{};
 
   suiteSetup(async () => await setupTestSuite(originalValues));
